refactor(acl,fwstate): use FQN versioned proto packages
Migrate the coupled acl and fwstate modules to fully-qualified, versioned
proto packages (modules.acl.controlplane.aclpb.v1 and
modules.fwstate.controlplane.fwstatepb.v1) so shared services such as
MetricsService and a future readiness service no longer collide on the
gateway. acl registers both services, so they migrate together. Updates the
Go control plane, both Rust CLIs and both web clients, and enables
buf/protolint enforcement.
