+++ .github/workflows/proto-lint.yml
@@ -46,7 +46,5 @@ jobs:
                                                 --exclude controlplane \
                                                 --exclude devices \
                                                 --exclude operators/bird-adapter \
-                                                --exclude modules/acl \
                                                 --exclude modules/balancer \
-                                                --exclude modules/fwstate \
                                                 --exclude modules/route-mpls
