+++ modules/acl/cli/src/main.rs
@@ -27,7 +27,7 @@ use ::commonpb::pb as commonpb;
 
 #[allow(non_snake_case)]
 pub mod aclpb {
-    tonic::include_proto!("aclpb");
+    tonic::include_proto!("modules.acl.controlplane.aclpb.v1");
 }
 
 #[derive(Serialize)]
