+++ modules/acl/controlplane/mod.go
@@ -7,15 +7,15 @@ import (
 	"google.golang.org/grpc"
 
 	"github.com/yanet-platform/yanet2/controlplane/ffi"
-	"github.com/yanet-platform/yanet2/modules/acl/controlplane/aclpb"
+	"github.com/yanet-platform/yanet2/modules/acl/controlplane/aclpb/v1"
 	fwstate "github.com/yanet-platform/yanet2/modules/fwstate/controlplane"
-	"github.com/yanet-platform/yanet2/modules/fwstate/controlplane/fwstatepb"
+	"github.com/yanet-platform/yanet2/modules/fwstate/controlplane/fwstatepb/v1"
 )
 
 const (
 	agentName          = "acl"
-	serviceName        = "aclpb.ACLService"
-	fwstateServiceName = "fwstatepb.FWStateService"
+	serviceName        = "modules.acl.controlplane.aclpb.v1.ACLService"
+	fwstateServiceName = "modules.fwstate.controlplane.fwstatepb.v1.FWStateService"
 )
 
 // ACLModule is a control-plane component for ACL (Access Control List) module
