+++ modules/acl/controlplane/service_test.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/yanet-platform/yanet2/controlplane/ffi"
 	"github.com/yanet-platform/yanet2/modules/acl/bindings/go/cacl"
-	"github.com/yanet-platform/yanet2/modules/acl/controlplane/aclpb"
+	"github.com/yanet-platform/yanet2/modules/acl/controlplane/aclpb/v1"
 )
 
 // fakeHandle is an in-memory implementation of ModuleHandle for tests.
